==> notebooks/0_dowload_preprocss_data.ipynb <==
# Packages 

%load_ext autoreload
%autoreload 2
import sys
import logging
import pandas as pd
base_dir = '../'
sys.path.append(base_dir)
import os
import numpy as np
import polars as pl
from src.config import raw_data_dir, raw_data_session_id_dir# task1_train_eval_dir

# Config 

# Download data

!mkdir ../{raw_data_dir}
!mkdir ../{raw_data_session_id_dir}
!mkdir ../{task1_train_eval_dir}

!aicrowd login

!aicrowd dataset list --challenge task-1-next-product-recommendation 
# ../{raw_data_dir}

!aicrowd dataset download --challenge task-1-next-product-recommendation

raw_data_dir

! ls {raw_data_dir}

# Preprocess data 

## Test data from phase 1 and 2 

task1_phase1 = pl.scan_csv(os.path.join('..', raw_data_dir, 'sessions_test_task1_phase1.csv'))
task1_phase2 = pl.scan_csv(os.path.join('..', raw_data_dir, 'sessions_test_task1.csv'))

print(task1_phase1.collect().shape)
print(task1_phase2.collect().shape)

print(task1_phase1.collect().head())
print(task1_phase2.collect().head())

print(task1_phase1.select('locale').collect().to_series().value_counts())
print(task1_phase2.select('locale').collect().to_series().value_counts())

task3_pl = pl.scan_csv(os.path.join('..', raw_data_dir, 'sessions_test_task3.csv'))

# task3_pl.head().collect()

# adding session_id to all data

train_pl = pl.scan_csv(os.path.join('..', raw_data_dir, 'sessions_train.csv'))
task1_pl = pl.scan_csv(os.path.join('..', raw_data_dir, 'sessions_test_task1.csv'))
task2_pl = pl.scan_csv(os.path.join('..', raw_data_dir, 'sessions_test_task2.csv'))
task3_pl = pl.scan_csv(os.path.join('..', raw_data_dir, 'sessions_test_task3.csv'))
product_pl = pl.scan_csv(os.path.join('..', raw_data_dir, 'products_train.csv'))

length_lst = []
for df in [train_pl, task1_pl, task2_pl, task3_pl, product_pl]:
    print()
    print()
    print(df.schema)
    shape = df.collect().shape
    length_lst.append(shape[0])
    print(shape)
    print()
    print(df.head(3).collect())
    print()
    print()

length_lst

train_pl = train_pl.with_columns(
    pl.Series(name='session_id', values=range(length_lst[0]))
    , pl.Series(name='random', values=np.random.uniform(size=length_lst[0]))
)
task1_pl= task1_pl.with_columns(
    pl.Series(name='session_id', values=range(length_lst[0],sum(length_lst[:2])))
)
task2_pl= task2_pl.with_columns(
    pl.Series(name='session_id', values=range(sum(length_lst[:2]),sum(length_lst[:3])))
)
task3_pl= task3_pl.with_columns(
    pl.Series(name='session_id', values=range(sum(length_lst[:3]),sum(length_lst[:4])))
)

# assert train_pl.select('session_id').max().collect()<200

train_pl.select('session_id').max().collect()

task1_pl.select('session_id').min().collect()

task2_pl.select('session_id').min().collect()

task3_pl.select('session_id').min().collect()

## Save data with session_id 

eval_df = train_pl.filter(pl.col('random')<=0.1).select(pl.all().exclude('random'))
train_df = train_pl.filter(pl.col('random')>0.1).select(pl.all().exclude('random'))

eval_df.collect().shape

train_df.collect().shape

train_pl.collect().shape

train_df.collect().write_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_train.parquet'))
eval_df.collect().write_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_eval.parquet'))


task1_pl.collect().write_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_test_task1.parquet'))
task2_pl.collect().write_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_test_task2.parquet'))
task3_pl.collect().write_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_test_task3.parquet'))

## Production data 

value_counts = product_pl.with_columns(pl.count().over('id').alias('cnt')).select(['id', 'cnt']).unique().collect().to_pandas()['cnt'].value_counts()

value_counts

value_counts/value_counts.sum()

**92% of products only in a single country; the rests are in more than one country**

### A example 

product_pl.with_columns(pl.count().over('id').alias('cnt')).sort(['cnt'], descending=True).head().collect()

product_pl.filter(pl.col('id')=='B0B9W7W8JD').collect()

### Save product info 

product_pl = product_pl.with_columns(
    pl.Series(name='unique_id', values=range(length_lst[-1]))
)

product_pl.schema

product_pl.collect().write_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'products_train.parquet'))


product_df = product_pl.select(['unique_id', 'id']).collect().to_pandas()
product_unique2id = dict(zip(product_df['unique_id'].values, product_df['id'].values))
with open(os.path.join(base_dir, raw_data_session_id_dir, 'product_unique2id.json'), 'wb') as f:
    pickle.dump(product_unique2id, f)

# assign session_id to test data of phase2

raw_data_session_id_dir

train_df = pl.scan_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_train.parquet'))
eval_df = pl.scan_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_eval.parquet'))
task1_pl = pl.scan_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_test_task1_phase1.parquet'))
task2_pl = pl.scan_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_test_task2_phase1.parquet'))
task3_pl = pl.scan_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_test_task3_phase1.parquet'))

length_lst = []
min_max_id = []
for df in [train_df, eval_df, task1_pl, task2_pl, task3_pl,# product_pl
          ]:
    # print()
    # print()
    # print(df.schema)
    min_id = df.select('session_id').min().collect()
    max_id = df.select('session_id').max().collect()
    shape = df.collect().shape
    length_lst.append(shape[0])
    min_max_id.append([min_id, max_id])
    # print(shape)
    # print()
    # print(df.head(3).collect())
    # print()
    # print()

length_lst

train_df.head().collect()

eval_df.filter(pl.col('session_id')==3273870).collect()

# min_max_id

3272918    ┆ null                 ┆ 0       │
│ 3273167    ┆ null                 ┆ 0       │
│ 3273760    ┆ null                 ┆ 0       │
│ 3273870    ┆ null                 ┆ 0       │
│ 3273928 

# min_max_id

max_id = 4014333


assert sum(length_lst) == max_id+1

task1_phase2 = pl.scan_csv(os.path.join('..', raw_data_dir, 'sessions_test_task1.csv'))
task2_phase2 = pl.from_pandas(pd.read_csv(os.path.join('..', raw_data_dir, 'sessions_test_task2.csv'))).lazy()
task3_phase2 = pl.from_pandas(pd.read_csv(os.path.join('..', raw_data_dir, 'sessions_test_task3.csv'))).lazy()
# product_ = pl.scan_csv(os.path.join('..', raw_data_dir, 'products_train.csv'))

assert task1_phase2.filter(pl.col('locale').is_null()).collect().shape[0] == 0

assert task2_phase2.filter(pl.col('locale').is_null()).collect().shape[0] == 0

assert task3_phase2.filter(pl.col('locale').is_null()).collect().shape[0] == 0

phase2_length_lst = []
for df in [task1_phase2, task2_phase2, task3_phase2]:
    print()
    print()
    print(df.schema)
    shape = df.collect().shape
    phase2_length_lst.append(shape[0])
    print(shape)
    print()
    print(df.head(3).collect())
    print()
    print()

max_id

phase2_length_lst

# phase2_l?ength_lst[:2]

task1_phase2= task1_phase2.with_columns(
    pl.Series(name='session_id', values=range(max_id+1,sum(phase2_length_lst[:1])+max_id+1))
)
task2_phase2= task2_phase2.with_columns(
    pl.Series(name='session_id', values=range(sum(phase2_length_lst[:1])+max_id+1,sum(phase2_length_lst[:2])+max_id+1))
)
task3_phase2= task3_phase2.with_columns(
    pl.Series(name='session_id', values=range(sum(phase2_length_lst[:2])+max_id+1,sum(phase2_length_lst[:3])+max_id+1))
)

# length_lst = []
# min_max_id = []
# for df in [train_df, eval_df, task1_pl, task2_pl, task3_pl,# product_pl
#           ]:
#     # print()
#     # print()
#     # print(df.schema)
#     min_id = df.select('session_id').min().collect()
#     max_id = df.select('session_id').max().collect()

    
# phase2_length_lst = []
for df in [task1_phase2, task2_phase2, task3_phase2]:
    min_id = df.select('session_id').min().collect()
    max_id = df.select('session_id').max().collect()
    print(f"[{min_id}, {max_id}]")
    # print()
    # print()
    # print(df.schema)
    # shape = df.collect().shape
    # phase2_length_lst.append(shape[0])
    # print(shape)
    # print()
    # print(df.head(3).collect())
    # print()
    # print()


task1_phase2.collect().write_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_test_task1.parquet'))
task2_phase2.collect().write_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_test_task2.parquet'))
task3_phase2.collect().write_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_test_task3.parquet'))

# Split train data to 2 parts 

train_pl = pl.scan_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_train.parquet'))


train_pl.head().collect()

length = train_pl.collect().shape[0]
length

train_pl.select('locale').collect().to_series().value_counts()

length 

train_pl = train_pl.with_columns(
    pl.Series(name='random', values=np.random.uniform(size=length))
)

train1 = train_pl.filter(pl.col('random')<=0.5).select(pl.all().exclude('random'))
train2 = train_pl.filter(pl.col('random')>0.5).select(pl.all().exclude('random'))

assert train1.collect().shape[0] + train2.collect().shape[0] == length

train1.collect().write_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_train1.parquet'))


train2.collect().write_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_train2.parquet'))
