==> notebooks/0_playground.ipynb <==
# Package 

%load_ext autoreload
%autoreload 2
import sys
import logging
base_dir = '../'
sys.path.append(base_dir)
import os
from gensim.similarities.annoy import AnnoyIndexer

from utils import *

import os
import pickle
import numpy as np
import pandas as pd
from gensim.models import Word2Vec
from tqdm import tqdm
import polars as pl
from annoy import AnnoyIndex
import polars as pl
import polars as pl
from utils import *
from src.eval import model_eval
from src.config import (
    raw_data_session_id_dir, candidate_dir, model_for_eval, candidate_file_name, submit_file_name
    # , w2v_model_file_name, w2v_index_file_name
)


# Config 

debug = False

topn = 100
model_version = 'w2v_v3'
if debug:
    n_rows = 100
else:
    n_rows = None

    
task = 'task1'
# Common setting
model_dir = f'../model_training/{model_version}/'
# target locales: locales needed for task1
target_locals = ["DE", 'JP', 'UK']
submit_file = os.path.join('../data/sub_files/', 
                           submit_file_name.format(
                                task=task
                               , model_version=model_version
                               , model_for_eval=model_for_eval
                               , topn=topn
                           )
                          )

train_cg_file = os.path.join(base_dir,
                             candidate_dir, 
            candidate_file_name.format(
                task=task
                , data_type='train'
                , model_version=model_version
                , model_for_eval=model_for_eval
                , topn=topn
            )
                            )
eval_cg_file = os.path.join(base_dir,
                            candidate_dir, 
            candidate_file_name.format(
                task=task
                , data_type='eval'
                , model_version=model_version
                , model_for_eval=model_for_eval
                , topn=topn
            )
                            )
test_cg_file = os.path.join(base_dir,
                            candidate_dir, 
            candidate_file_name.format(
                task=task
                , data_type='test'
                , model_version=model_version
                , model_for_eval=model_for_eval
                , topn=topn
            )
                            )

# specific setting
num_tree = 100
# if model_for_eval:
w2v_model_file = os.path.join(model_dir, f"{model_for_eval}.model")
annoy_index_file = os.path.join(model_dir, f"{str(num_tree)}_{model_for_eval}.index")

# Load model 

w2vec = Word2Vec.load(w2v_model_dir)

test_sessions = read_test_data(task, test_data_dir=test_data_dir)
test_sessions.sample(5)

test_sessions.info()

# test_sessions['prev_items'] = test_sessions.apply(lambda row: process_item_lst(row), axis=1)

df = test_sessions.sample(1000)

for col in df.columns:
    df[col] = df[col].astype(str)

df.info()

def get_rec(prev_items):
    # print(prev_items)
    res = [ele.replace('[', '').replace(']', '').replace('\n', '').replace("'", '').replace(' ', '') for ele in prev_items.split(' ')]
    
    try:
        similarity_dic = w2vec.wv.most_similar(positive=prev_items, topn=100)
        res = [item for item, simi in similarity_dic] 
        # print(err)
    except Exception as e:
        print(e)
    
    return res

pl_df = pl.from_dataframe(df)
pl_df = (
    pl_df
        .with_column(pl.col('prev_items').apply(lambda row: get_rec(row)).alias('next_item_prediction'))
)
result_df = pl_df.to_pandas()

# pl_df

result_df.head()

# Polars pl.element() 

raw_data_session_id_dir

! ls ../data/raw_data_session_id

target_pl = pl.scan_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_train.parquet'),
                          n_rows=n_rows).filter(pl.col('locale').is_in(target_locals)).with_columns(pl.col('prev_items').apply(str2list))
file_path = '../model_training/als_v1/itemid2item.json'
with open(file_path, 'rb') as f:
    itemid2item = pickle.load(f)




item2itemid = {value: key for key, value in itemid2item.items()}

# item2itemid

target_pl.head().collect()

target_pl.schema

# dir(pl.element())

# item2itemid['B09W9FND7K']

%load_ext memory_profiler

%%time
%memit

def map_item(lst):
    item_ids  = [item2itemid[item] for item in lst]
    return item_ids

df1 = (
    target_pl.with_columns(
                pl.col('prev_items').apply(lambda x: map_item(x)).alias('prev_item_ids')
    ).select(
        ['session_id', 'prev_items', 'prev_item_ids']
    )
).collect()
# print(df1.shape)
# print(df1.head(2))

%%time
%memit

df2 = (
    target_pl.explode('prev_items')
        .with_columns(
                pl.col('prev_items').map_dict(item2itemid).alias('prev_item_ids'))
        .groupby('session_id')
        .agg(
            pl.col('prev_items')
            , pl.col('prev_item_ids')
        )
).collect()
print(df2.shape)
print(df2.head(2))

# W2V in details 

eval_pl = pl.scan_parquet('../data/candidates/task1_eval_w2v_v3_True_top100.parquet')
original_eval_pl = pl.scan_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_eval.parquet'), n_rows=n_rows).filter(pl.col('locale').is_in(target_locals)).with_columns(pl.col('prev_items').apply(str2list))


eval_pl.head().collect()

original_eval_pl.head().collect()

merged_pl = original_eval_pl.join(eval_pl, how='left', on='session_id' )

merged_pl.head().collect()

# def get_next_items(x):
#     prev_items = x['prev_items']
#     local_rec = x['next_item_prediction']
#     final = [ele for ele in local_rec if ele in prev_items]
#     return len(final)

# (merged_pl
#     .with_columns(
        
#     pl.struct(["prev_items", "next_item_prediction"]).apply(
#                         lambda x: get_next_items(x)).alias('duplicate_num'))
#     # .select(pl.col('duplicate_num')).collect().describe()
# ).head().collect()
