==> notebooks/1_task1_ALS_v2.ipynb <==
# Packages 

%load_ext autoreload
%autoreload 2
import sys
import logging
base_dir = '../'

sys.path.append(base_dir)
import os
from utils import *

import os
import numpy as np
import pandas as pd
from gensim.models import Word2Vec
import pickle


from annoy import AnnoyIndex
import polars as pl
import implicit
import scipy.sparse as sps
from src.eval import get_recall_at_k, pd_get_recall_at_k
from src.config import raw_data_session_id_dir, candidate_dir, model_for_eval
from src.eval import model_eval


# Config 

debug = False

model_version = 'als_v2'

submit_res = False

topn = 100
if debug:
    n_rows = 1000
else:
    n_rows = None
train_data_dir = '.'
test_data_dir = '.'
task = 'task1'

model_dir = f'../model_training/{model_version}'

model_file = os.path.join(model_dir, 'als_model.pkl')

target_locals = ["DE", 'JP', 'UK']


sub_file = f'../data/sub_files/{model_version}.parque'

# sub_file

! mkdir {model_dir}

model_file

# Function 

def get_rec(target_pl, model, user_item):
    user_lst = target_pl.select('session_id').collect().to_series().to_list()
    topn = 100

    next_items, item_similarity = model.recommend(user_lst, user_item[user_lst, :],
                                  filter_already_liked_items=True,
                                   N=topn
                                 )#[0].tolist()
    from tqdm import tqdm

    final_items = []
    for lst in tqdm(next_items):
        final_items.append([product_unique2id[ele] for ele in lst])
    eval_res = (
        target_pl.with_columns(
            pl.Series(name='next_item_prediction', values=final_items)
            , pl.Series(name='als_similarity', values=item_similarity)
            )
            # .with_columns(
            #     # pl.col("next_item_prediction").arr.eval(
            #     #     # product_unique2id[pl.element()]
            #     #     # dir(pl.element())
            #     #     # pl.element().cast(pl.Int8)#.map(product_unique2id)
            #     #                                        ).alias('test')
            # )
    )
    return eval_res

# Read data

train_pl = pl.scan_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_train.parquet'), n_rows=n_rows).filter(pl.col('locale').is_in(target_locals)).with_columns(pl.col('prev_items').apply(str2list))

eval_pl = pl.scan_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_eval.parquet'), n_rows=n_rows).filter(pl.col('locale').is_in(target_locals)).with_columns(pl.col('prev_items').apply(str2list))

# df_sess.head(3).collect()
test_pl = pl.scan_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_test_task1.parquet'), n_rows=n_rows).with_columns(pl.col('prev_items').apply(str2list))
product_pl = pl.scan_parquet(
    os.path.join(base_dir, raw_data_session_id_dir, 'products_train.parquet'), 
)
with open(os.path.join(base_dir, raw_data_session_id_dir, 'product_unique2id.json'), 'rb') as f:
    product_unique2id  = pickle.load(f)

train_data = train_pl.with_columns(
    pl.col('prev_items').arr.concat(pl.col('next_item')) 
)
if not model_for_eval:
    eval_data = eval_pl.with_columns(
    pl.col('prev_items').arr.concat(pl.col('next_item')) 
)
else:
    eval_data = eval_pl
test_data = test_pl

# Process Training data 

cols_to_keep = ['prev_items', 'locale','session_id']
user_item_pl = (
    pl.concat([train_data.select(cols_to_keep), 
               eval_data.select(cols_to_keep), 
               test_data.select(cols_to_keep)], how='vertical')
        .explode('prev_items')
        .join(product_pl.select(['id', 'locale', 'unique_id']), how='left', 
              left_on=['prev_items', 'locale']
              , right_on=['id', 'locale']
             )
        .select(
            pl.col('session_id')
            , pl.col('unique_id').alias('item_id')
        )
)

user_item = user_item_pl.collect().to_pandas()

user_item.shape

user_item.head()

# ALS Model

## Model config 

vec_size = 100
iterations = 20
if debug:
    iterations = 2

# user_item['session_id']

## training data 

user_item = sps.coo_matrix(
      (
          np.ones(user_item.shape[0]), # We're using a matrix of ones, but using type weights or repurchase weights could help!
          (user_item['session_id'],
          user_item['item_id'])
      ),
      dtype='int8'
    ).tocsr()


user_item.shape


## Model Training 

%%time
model = implicit.als.AlternatingLeastSquares(factors=vec_size, regularization=0.1, iterations=iterations)

model.fit(user_items=user_item, show_progress=True)

## Save model 

debug

model_file

user_item

# if not debug:
sps.save_npz(os.path.join(model_dir, "user_item.npz"), user_item)
with open(model_file, 'wb') as f:
    pickle.dump(model, f)
del model
del user_item

! ls {model_dir}

# Load model

user_item = sps.load_npz(os.path.join(model_dir, "user_item.npz"))
with open(model_file, 'rb') as f:
    model = pickle.load(f)

# Model eval 

%%time

eval_res = get_rec(target_pl=eval_pl, model=model, user_item=user_item)

eval_res.schema

%%time
model_eval(eval_res)

# submit res

%%time
test_res = get_rec(target_pl=test_pl, model=model, user_item=user_item)

predictions = test_res.collect().to_pandas()

check_predictions(predictions, test_sessions=test_pl.collect().to_pandas(), 
                  # check_products=True, product_df=products
                 )
# Its important that the parquet file you submit is saved with pyarrow backend
if not debug:
    predictions.to_parquet(sub_file, engine='pyarrow')

# # You can submit with aicrowd-cli, or upload manually on the challenge page.
!aicrowd submission create -c task-1-next-product-recommendation -f {sub_file}

test_model = implicit.als.AlternatingLeastSquares(factors=50)

help(test_model.fit)
