==> notebooks/1_task1_next_item_counter_v2_for_eval.ipynb <==
# Package 

%load_ext autoreload
%autoreload 2
import sys
import logging
base_dir = '../'
sys.path.append(base_dir)
import os
import warnings
warnings.simplefilter('ignore')

import pickle
import gc
import re
import polars as pl
from collections import defaultdict, Counter

import numpy as np
import pandas as pd
pd.set_option('display.max_columns', None)
from tqdm.auto import tqdm
import polars as pl
from utils import *
from src.eval import model_eval
from src.config import raw_data_session_id_dir, candidate_dir, model_for_eval

# model_for_eval

# df_prod = pd.read_csv('data/products_train.csv')
# df_prod

# Config 

debug = False

model_version = 'next_item_counter_v2'

topn = 100
if debug:
    n_rows = 1000
else:
    n_rows = None
# debug_session_num = 100
train_data_dir = '.'
test_data_dir = '.'
task = 'task1'

model_dir = f'../model_training/{model_version}/'

# target locales: locales needed for task1
target_locals = ["DE", 'JP', 'UK']

# if model_for_eval:
model_file = os.path.join(model_dir, f'nic_{model_for_eval}_for_eval.parquet')
submit_file = os.path.join('../data/sub_files/', f'submission_{task}_nic_{model_for_eval}_for_eval.parquet')

candidate_file.format(
    task='task1'
    , data_type='train'
    , model_version=model_version
    , model_for_eval=model_for_eval
    , topn=topn
)

! mkdir {model_dir}

model_file

submit_file

# Data 

! ls ../{raw_data_session_id_dir}

train_pl = pl.scan_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_train.parquet'), n_rows=n_rows).filter(pl.col('locale').is_in(target_locals)).with_columns(pl.col('prev_items').apply(str2list))

eval_pl = pl.scan_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_eval.parquet'), n_rows=n_rows).filter(pl.col('locale').is_in(target_locals)).with_columns(pl.col('prev_items').apply(str2list))

# df_sess.head(3).collect()
test_pl = pl.scan_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_test_task1.parquet'), n_rows=n_rows).with_columns(pl.col('prev_items').apply(str2list))


# Function 

# 'item', 'next_item_prediction', 'next_item_weight'

def nic_rec(target_pl, nic_model):
    rec_num = 100
    final_cols = ['session_id', 'next_item_prediction', 'rec_num']
    target_pl = (
        target_pl
            .with_columns(
                pl.col('prev_items').arr.get(-1).alias('last_item')
            )
            .join(nic_model, how='left', left_on='last_item', right_on='item')
            .with_columns(
                pl.when(pl.col('next_item_prediction').is_null()).then([]).otherwise(pl.col('next_item_prediction').arr.head(rec_num)).alias('next_item_prediction')
            )
            .with_columns(
                pl.col('next_item_prediction').arr.lengths().alias('rec_num')
            )
            .select(
                final_cols
            )
    )#.head(3).collect()
    return target_pl

# Next Item Statistics 

train_data = train_pl.with_columns(
    pl.col('prev_items').arr.concat(pl.col('next_item')) 
)
if not model_for_eval:
    eval_data = eval_pl.with_columns(
    pl.col('prev_items').arr.concat(pl.col('next_item')) 
)
else:
    eval_data = eval_pl
test_data = test_pl

# model_for_eval

# if not model_for_eval:


cols_to_keep = ['prev_items']
next_items_pl = (
    pl.concat([train_data.select(cols_to_keep), eval_data.select(cols_to_keep), test_data.select(cols_to_keep)], how='vertical')
        .with_columns(
            pl.col('prev_items').arr.shift(-1).alias('next_item_lst')
            , pl.col('prev_items').arr.lengths().alias('length')
        )
        .select(
            pl.col('prev_items').arr.head(pl.col('length')-1).alias('prev')
            , pl.col('next_item_lst').arr.head(pl.col('length')-1).alias('next')
        )
        .explode(['prev','next' ])
        .groupby(['prev','next' ])
        .agg(
            pl.count().alias('cnt')
        )
        .sort(['prev', 'cnt'], descending=True)
        .with_columns(
            pl.col('cnt').max().over('prev').alias('max_count')
            , pl.col('cnt').min().over('prev').alias('min_count')
        )
        .with_columns(
            pl.when(pl.col('max_count')==pl.col('min_count')).then(1).otherwise((pl.col('cnt')-pl.col('min_count'))/(pl.col('max_count')-pl.col('min_count'))).alias('normalized_cnt')
        )
        .groupby('prev')
        .agg(
            pl.col('next').alias('next_item_prediction')
            , pl.col('cnt').alias('next_item_cnt')
            , (pl.col('normalized_cnt')+pl.lit(0.01)).alias('next_item_weight')
        )
        .select(
            pl.col('prev').alias('item')
            , 'next_item_prediction'
            , pl.col('next_item_weight')
        )
        
)

# next_items_pl.collect()#.filter(pl.col('item')=='B09LXX1PQ9')

# next_item_dict = defaultdict(list)

# for _, row in tqdm(df_sess.iterrows(), total=len(df_sess)):
#     prev_items = str2list(row['prev_items'])
#     if not model_for_eval:
#         next_item = row['next_item']
#     prev_items_length = len(prev_items)
#     if prev_items_length <= 1:
#         if not model_for_eval:
#             next_item_dict[prev_items[0]].append(next_item)
#     else:
#         for i, item in enumerate(prev_items[:-1]):
#             next_item_dict[item].append(prev_items[i+1])
#         if not model_for_eval:
#             next_item_dict[prev_items[-1]].append(next_item)

# next_item_dict

# for _, row in tqdm(df_test.iterrows(), total=len(df_test)):
#     prev_items = str2list(row['prev_items'])
#     prev_items_length = len(prev_items)
#     if prev_items_length <= 1:
#         continue
#     else:
#         for i, item in enumerate(prev_items[:-1]):
#             next_item_dict[item].append(prev_items[i+1])

# # next_item_map = {}
# topn = 100
# item_lst = []
# common_items_lst = []
# weights_lst = []
# for item in tqdm(next_item_dict):
#     counter = Counter(next_item_dict[item])
#     most_common_cnt = counter.most_common(1)[0][1]
#     most_common_lst = list(zip(*counter.most_common(topn)))
#     most_common_lst[1] = list(np.array(most_common_lst[1])/most_common_cnt)
#     item_lst.append(item)
#     common_items_lst.append(list(most_common_lst[0]))
#     weights_lst.append(most_common_lst[1])
#     # next_item_map[item] = most_common_lst
#     # next_item_map[item] = [i[0] for i in counter.most_common(100)]

# next_item_df = pd.DataFrame(
#     {'item': item_lst
#     , 'next_item_prediction': common_items_lst
#      , 'next_item_weight': weights_lst
#     }
# )

# next_item_df.sample(10)

# next_items_pl.collect().filter(pl.col('item')=='B07QGW8LFT')

# nex_item_pl = pl.from_pandas(next_item_df).lazy().select(
#     'item'
#     , pl.col('next_item_prediction').alias('next_item_rec')
#     , 'next_item_weight'
# )

# with open('../model_training/next_item_counter/model.pkl', 'rb') as f:
#     model = pickle.load(f)

# model.keys()

# model['next_item_map']

# nex_item_pl = pl.DataFrame(
#     {
#         'item': model['next_item_map'].keys()
#         , 'next_item_rec': model['next_item_map'].values()
#     }
# ).lazy()

next_items_pl.collect().shape

## Save model 

next_items_pl.collect().write_parquet(model_file)

del next_items_pl

## Read Model 

model_file

next_items_pl = pl.scan_parquet(model_file)

type_dict = next_items_pl.schema

type_dict.keys()

## Model eval 

# train_pl.schema

# eval_pl.schema

eval_candidate_pl = nic_rec(target_pl=eval_pl, nic_model=next_items_pl)#.head().collect()


# if_hit = pl.element().rank()
target_df = eval_pl.join(eval_candidate_pl, how='left', on='session_id')


# eval_final.head().collect()

# target_df.schema

model_eval(target_df=target_df)

## Candidate Saving 

train_candidate_pl = nic_rec(target_pl=train_pl, nic_model=next_items_pl)#.head().collect()

# eval_candidate_pl = nic_rec(target_pl=eval_pl, nic_model=next_items_pl)#.head().collect()

test_candidate_pl = nic_rec(target_pl=test_pl, nic_model=next_items_pl)#.head().collect()

train_candidate_pl.collect().write_parquet(os.path.join(base_dir, candidate_dir, 'task1_train_nic_top100.parquet'))

eval_candidate_pl.collect().write_parquet(os.path.join(base_dir, candidate_dir, 'task1_eval_nic_top100.parquet'))

test_candidate_pl.collect().write_parquet(os.path.join(base_dir, candidate_dir, 'task1_test_nic_top100.parquet'))

## Save test result

predictions = test_pl.join(test_candidate_pl, how='left', on='session_id').collect()[['locale', 'next_item_prediction']].to_pandas()

submit_file

check_predictions(predictions, test_sessions=test_pl.collect().to_pandas(), 
                  # check_products=True, product_df=products
                 )
# Its important that the parquet file you submit is saved with pyarrow backend
predictions.to_parquet(submit_file, engine='pyarrow')

!aicrowd submission create -c task-1-next-product-recommendation -f {submit_file}

# Top200 for fallback logics

# next_item_df.head()

# next_item_df[cols].info()

popular_df = pd.concat([df_sess[['prev_items', 'locale']], df_test[['prev_items', 'locale']]], axis=0)

popular_df.shape

popular_pl = pl.from_pandas(popular_df).lazy()

topn = 200
locale_popular_pl = (
    popular_pl
        .select(
            pl.col('prev_items').apply(str2list)#.explode().alias('item')
            , pl.col('locale')
        )
        .explode('prev_items')#.alias('item')
        .groupby(['locale', 'prev_items'])
        .agg(
            pl.count()
        )
        .with_columns(
            pl.col('count').rank(method='ordinal', descending=True).over('locale').alias('rank')
        )
        .filter(pl.col('rank')<=topn)
        .with_columns(
            pl.col('count').max().over('locale').alias('max_count')
            , pl.col('count').min().over('locale').alias('min_count')
        )
        .with_columns(
            ((pl.col('count')-pl.col('min_count'))/(pl.col('max_count')-pl.col('min_count'))).alias('weight')
        )
        .sort('locale', 'rank')
        .select(
            'locale'
            , 'prev_items'
            , 'weight'
        )
        .groupby('locale')
        .agg(
            pl.col('weight').alias('locale_popular_weight')
            , pl.col('prev_items').alias('locale_popular_rec')
        )
        # .count()#.head(3).collect())
        # .collect()
)

# locale_popular_pl.collect()

# locale_popular_pl.schema

# popular_df.apply(lambda x: str2list(x['prev_items']), axis=1)

# df_sess.head()

# df_test.head()

# # next_item_df['next_item_prediction'] = next_item_df['next_item_prediction'].astype(str)
# # next_item_df['next_item_weights'] = next_item_df['next_item_weights'].astype(str)
# cols = [
#     # 'item',
#         'next_item_prediction'
#         , 'next_item_weights'
#        ]
# next_item_pl = pl.from_pandas(next_item_df[cols])

# next_item_pl

# # k = []
# # v = []

# # for item in next_item_dict:
# #     k.append(item)
# #     v.append(next_item_dict[item])
    
# # df_next = pd.DataFrame({'item': k, 'next_item': v})
# df_next = next_item_df.explode('next_item_prediction').reset_index(drop=True)
# df_next = df_next.merge(products, how='left', left_on='item', right_on='id')
# df_next

# df_next['next_item'].value_counts().index.tolist()[:200]

# model = {
#     'top200': top200
#     , 'next_item_map': next_item_map
# }

## Save model 

# model_file

# with open(model_file, 'wb') as f:
#     pickle.dump(model, f)

# Get final result 

## Load Model 

# # with open(model_file, 'rb') as f:
#     model = pickle.load(f)

# top200

# next_item_map

# def get_rec(target_df, model):
#     next_item_map = model['next_item_map']
#     top200  = model['top200']
#     target_df['last_item'] = target_df['prev_items'].apply(lambda x: str2list(x)[-1])
#     target_df['next_item_prediction'] = target_df['last_item'].map(next_item_map)
#     preds = []

#     for _, row in tqdm(target_df.iterrows(), total=len(target_df)):
#         pred_orig = row['next_item_prediction']
#         pred = pred_orig
#         prev_items = str2list(row['prev_items'])
#         if type(pred) == float:
#             pred = top200[:100]
#         else:
#             if len(pred_orig) < 100:
#                 for i in top200:
#                     if i not in pred_orig and i not in prev_items:
#                         pred.append(i)
#                     if len(pred) >= 100:
#                         break
#             else:
#                 pred = pred[:100]
#         preds.append(pred)
#     target_df['next_item_prediction'] = preds
#     print(target_df['next_item_prediction'].apply(len).describe())
#     return target_df

# model.keys()

# model['next_item_map']

# Candidate for train data 

# train_pl = pl.scan_csv('sessions_train.csv')
train_pl = pl.scan_parquet('../data/eval_data/next_item_counter_train_eval_300k.parquet')

target_locals

train_pl.schema

train_pl = (
    train_pl
        .filter(pl.col('locale').is_in(target_locals))
        .with_columns(
            pl.col('prev_items').apply(str2list).arr.get(-1).alias('last_item')
        )
        .join(nex_item_pl, how='left', left_on='last_item', right_on='item')
        .with_columns(
            pl.when(pl.col('next_item_rec').is_null()).then([]).otherwise(pl.col('next_item_rec').arr.head(100)).alias('next_item_prediction')
        )
        .with_columns(
            pl.col('next_item_prediction').arr.lengths().alias('rec_num')
        )
        .select(
            'prev_items'
            , 'next_item'
            , 'locale'
            , 'next_item_prediction'
            , 'rec_num'
        )
)#.head(2).collect()

train_pl.collect().write_parquet('../data/candidates/task1_train_nic_without_pupular_top100_300k.parquet')

# Final resul 

def pl_rec(target_pl, locale_popular_pl, nex_item_pl):
    rec_num = 100
    target_pl = (
        target_pl
            .with_columns(
                pl.col('prev_items').apply(str2list).arr.get(-1).alias('last_item')
            )
            .join(nex_item_pl, how='left', left_on='last_item', right_on='item')
            .join(locale_popular_pl, how='left', on='locale')
            .with_columns(
                pl.when(pl.col('next_item_rec').is_null()).then([]).otherwise(pl.col('next_item_rec')).alias('next_item_rec')
            )
            .with_columns(
                pl.concat_list([pl.col('next_item_rec'), pl.col('locale_popular_rec')])
                    .alias('next_item_prediction')
                    .arr.head(rec_num)

            )
            .with_columns(
                pl.col('next_item_prediction').arr.lengths().alias('rec_num')
            )
    )#.head(3).collect()
    return target_pl

eval_pl = pl.scan_parquet(f'../data/eval_data/w2v_train_eval_result_300k.parquet')

eval_pl.schema

nex_item_pl.schema

# locale_popular_pl.head(3).collect()

eval_pl = pl_rec(target_pl=eval_pl, locale_popular_pl=locale_popular_pl, nex_item_pl=nex_item_pl)

# eval_pl.head(3).collect()

eval_pl.select(
    pl.col('next_item_prediction').arr.head(20).arr.contains(pl.col('next_item')).mean().alias('recall@20')
    , pl.col('next_item_prediction').arr.head(100).arr.contains(pl.col('next_item')).mean().alias('recall@100')
).collect()


# eval_cols = ['len', 'recall@20', 'recall@100']
# train_eval_df[eval_cols] = train_eval_df.apply(pd_get_recall_at_k, axis=1, result_type='expand')
# print(train_eval_df[eval_cols].mean())

# train_eval_df.shape

# model_version

eval_pl.collect().shape

eval_pl.collect().write_parquet(f'../data/eval_data/{model_version}_train_eval_300k.parquet', 
                      # engine='pyarrow'
                     )

# Submit result 

test_pl = pl.scan_csv('sessions_test_task1.csv')
test_pl = pl_rec(target_pl=test_pl, locale_popular_pl=locale_popular_pl, nex_item_pl=nex_item_pl)

submit_file

! ls -al | grep {submit_file}

test_pl.collect().shape

submit_file

test_pl.head(3).collect()

# test_pl.collect().select('locale', 'next_item_prediction').write_parquet(submit_file,
#                                                                          # engine='pyarrow'
#                                                                         )

# # You can submit with aicrowd-cli, or upload manually on the challenge page.
# !aicrowd submission create -c task-1-next-product-recommendation -f {submit_file}

# Rank  

w2v_df = pl.scan_parquet('submission_task1.parquet')

assert w2v_df.collect().shape[0] == test_pl.collect().shape[0]

w2v_df.schema

# test_pl.head()

target_df = pl.concat([test_pl.select('prev_items', 'locale', 'next_item_rec').collect(), w2v_df.select('next_item_prediction').collect()]
                    , how='horizontal' )

target_df.shape

target_df.head(10)

rec_num = 100
target_pl = (
    target_df
        .lazy()
        .select(
            'prev_items'
            , 'locale'
            , pl.concat_list([pl.col('next_item_rec'), pl.col('next_item_prediction')])
                .alias('next_item_prediction')
                .arr.head(rec_num)

        )
        .with_columns(
            pl.col('next_item_prediction').arr.lengths().alias('rec_num')
        )
)#.head(3).collect()

target_pl.schema

target_pl.head(6).collect()

# ! mkdir ../data/sub_files

target_pl.collect().select('locale', 'next_item_prediction').write_parquet('../data/sub_files/rank_v1.parquet',
                                                                         # engine='pyarrow'
                                                                        )

# # You can submit with aicrowd-cli, or upload manually on the challenge page.
# !aicrowd submission create -c task-1-next-product-recommendation -f '../data/sub_files/rank_v1.parquet'

## Rank2 

rec_num = 100
target_pl = (
    target_df
        .lazy()
        .select(
            'prev_items'
            , 'locale'
            , pl.concat_list([pl.col('next_item_rec').arr.head(20), pl.col('next_item_prediction')])
                .alias('next_item_prediction')
                .arr.head(rec_num)

        )
        .with_columns(
            pl.col('next_item_prediction').arr.lengths().alias('rec_num')
        )
)#.head(3).collect()

target_pl.collect().select('locale', 'next_item_prediction').write_parquet('../data/sub_files/rank_v2.parquet',
                                                                         # engine='pyarrow'
                                                                        )

# # You can submit with aicrowd-cli, or upload manually on the challenge page.
!aicrowd submission create -c task-1-next-product-recommendation -f '../data/sub_files/rank_v2.parquet'

# Debug 

test_pl = pl.scan_parquet(submit_file)

test_pl.schema

test_pl.head(5).collect()

test_pl.select(
    pl.col('next_item_prediction').arr.lengths().min()
    , pl.col('next_item_prediction').arr.lengths().max()
).collect()
