==> notebooks/1_task1_original_1_task1_ALS.ipynb <==
# Packages 

%load_ext autoreload
%autoreload 2
import sys
import logging
base_dir = '../'

sys.path.append(base_dir)
import os
from utils import *

import os
import numpy as np
import pandas as pd
from gensim.models import Word2Vec
import pickle

from annoy import AnnoyIndex
import polars as pl
import implicit
import scipy.sparse as sps
from src.eval import get_recall_at_k, pd_get_recall_at_k, model_eval
from src.config import raw_data_session_id_dir, candidate_dir, model_for_eval

# Config 

debug = False

model_version = 'als_v1'

topn = 100
if debug:
    n_rows = 1000
else:
    n_rows = None
train_data_dir = '.'
test_data_dir = '.'
task = 'task1'

model_for_eval = True

model_dir = f'../model_training/{model_version}'

model_file = os.path.join(model_dir, 'als_model.pkl')

target_locals = ["DE", 'JP', 'UK']

submit_file = os.path.join('../data/sub_files/', 'als_v1.parquet')


! mkdir {model_dir}

# model_file

# Read data

# train_pl = pl.scan_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_train.parquet'), n_rows=n_rows).filter(pl.col('locale').is_in(target_locals)).with_columns(pl.col('prev_items').apply(str2list))

# eval_pl = pl.scan_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_eval.parquet'), n_rows=n_rows).filter(pl.col('locale').is_in(target_locals)).with_columns(pl.col('prev_items').apply(str2list))

# # df_sess.head(3).collect()
# test_pl = pl.scan_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_test_task1.parquet'), n_rows=n_rows).with_columns(pl.col('prev_items').apply(str2list))


# train_data = train_pl.with_columns(
#     pl.col('prev_items').arr.concat(pl.col('next_item')) 
# )
# if not model_for_eval:
#     eval_data = eval_pl.with_columns(
#     pl.col('prev_items').arr.concat(pl.col('next_item')) 
# )
# else:
#     eval_data = eval_pl
# test_data = test_pl
# cols_to_keep = ['prev_items', 'locale']
# popular_pl = pl.concat([train_data.select(cols_to_keep), eval_data.select(cols_to_keep), test_data.select(cols_to_keep)], how='vertical')
        

# debug_session_num = 1000
# train_sessions = read_train_data(train_data_dir=train_data_dir)
# test_sessions = read_test_data(task, test_data_dir=test_data_dir)
# train_sessions = train_sessions[train_sessions['locale'].isin(target_locals)]

# if debug:
#     train_sessions = train_sessions.sample(debug_session_num)
#     test_sessions = test_sessions.sample(debug_session_num)
# f"{train_sessions.shape}; {test_sessions.shape}"

# with_columns(pl.col('prev_items').apply(str2list))

# str2list

train_pl = pl.scan_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_train.parquet'), n_rows=n_rows).filter(pl.col('locale').is_in(target_locals)).with_columns(pl.col('prev_items').apply(str2list)).collect()

eval_pl = pl.scan_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_eval.parquet'), n_rows=n_rows).filter(pl.col('locale').is_in(target_locals)).with_columns(pl.col('prev_items').apply(str2list)).collect()

# df_sess.head(3).collect()
test_pl = pl.scan_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_test_task1.parquet'), n_rows=n_rows).with_columns(pl.col('prev_items').apply(str2list)).collect()
product_pl = pl.scan_parquet(
    os.path.join(base_dir, raw_data_session_id_dir, 'products_train.parquet'), 
)
with open(os.path.join(base_dir, raw_data_session_id_dir, 'product_unique2id.json'), 'rb') as f:
    product_unique2id  = pickle.load(f)

train_sessions = train_pl.with_columns(
    pl.col('prev_items').arr.concat(pl.col('next_item')) 
).to_pandas()
if not model_for_eval:
    eval_sessions = eval_pl.with_columns(
    pl.col('prev_items').arr.concat(pl.col('next_item')) 
).to_pandas()
else:
    eval_sessions = eval_pl.to_pandas()
test_sessions = test_pl.to_pandas()

train_sessions['type'] = 'train'
eval_sessions['type'] = 'eval'
test_sessions['type'] = 'test'
train_sessions.reset_index(inplace=True)
eval_sessions.reset_index(inplace=True)
test_sessions.reset_index(inplace=True)

# train_sessions.head()

# test_sessions.head()

# train_sessions[cols_to_keep].head()

# train_df.head()

cols_to_keep = ['index', 'type', 'prev_items', 'locale']

original_train_test_df = pd.concat([train_sessions[cols_to_keep], eval_sessions[cols_to_keep], 
                                    test_sessions[cols_to_keep]], 
                          axis=0)

original_train_test_df.head()

train_test_pl = pl.from_pandas(original_train_test_df)

def process_prev_item(prev_items):
    # print(prev_items)
    res = [ele.replace('[', '').replace(']', '').replace('\n', '').replace("'", '').replace(' ', '') for ele in prev_items.split(' ')]
    return res

train_test_pl = (
    train_test_pl
        # .with_columns(pl.col('prev_items').apply(lambda row: process_prev_item(row)))
        .with_columns(pl.col('index').cast(pl.Utf8))
        .with_columns((pl.col('index')+pl.col('type')).alias('user'))
        # .with_columns(pl.col('prev_items').explode())
)

train_test_df = train_test_pl.to_pandas()

train_test_df.head()

train_test_df.head()

train_test_df['user'] = train_test_df['user'].astype("category")
train_test_df['user_id'] = train_test_df['user'].cat.codes

test_sessions['index'] = test_sessions['index'].astype(str)
test_sessions = test_sessions.merge(train_test_df[['index', 'type', 'user_id']], how='left', on=['index', 'type'])

train_sessions['index'] = train_sessions['index'].astype(str)
train_sessions = train_sessions.merge(train_test_df[['index', 'type', 'user_id']], how='left', on=['index', 'type'])

eval_sessions['index'] = eval_sessions['index'].astype(str)
eval_sessions = eval_sessions.merge(train_test_df[['index', 'type', 'user_id']], how='left', on=['index', 'type'])

train_test_df = train_test_df.explode('prev_items')
train_test_df.shape

test_sessions.head()

train_test_df.head()



train_test_df['prev_items'] = train_test_df['prev_items'].astype("category")
train_test_df['item_id'] = train_test_df['prev_items'].cat.codes

train_test_df.head()

len(train_test_df)

itemid2item = dict(zip(train_test_df['item_id'], train_test_df['prev_items']))

len(itemid2item)

max(itemid2item.keys())

# itemid2item = 
# train_test_df.groupby(['prev_items', 'item_id'])['user'].count().reset_index()

# itemid2item

# ALS Model

train_test_df.head()

# user_item[1243, :]

user_item = sps.coo_matrix(
      (
          np.ones(train_test_df.shape[0]), # We're using a matrix of ones, but using type weights or repurchase weights could help!
          (train_test_df['user_id'],
          train_test_df['item_id'])
      ),
      dtype='int8'
    ).tocsr()


user_item.shape


model = implicit.als.AlternatingLeastSquares(factors=100, regularization=0.1, iterations=20)

model.fit(user_items=user_item, show_progress=True)

# help(model.recommend_all)

# help(model.recommend)

def get_rec(user_id, model, user_item, itemid2item, topn=100):
    item_ids = model.recommend(user_id, 
                              user_item[user_id, :],
                              filter_already_liked_items=True,
                               N=topn
                             )[0].tolist()
    # print(item_ids)
    items = [itemid2item[key] for key in item_ids]
    return items
    

assert model.item_factors.shape[0] == len(itemid2item)

# itemid2item[7550]

# itemid2item


# help(model.recommend)

# original_train_test_df

# test_prediction = train_test_df[(train_test_df['type']=='test')]['item_id'].unique()

# test_sessions.shape

# predictions.head()

## Save model 

debug

model_dir

# sps.save_npz(os.path.join(model_dir, "user_item.npz"), user_item)
# with open(model_file, 'wb') as f:
#     pickle.dump(model, f)
# with open(os.path.join(model_dir, 'itemid2item.json'), 'wb') as f:
#     pickle.dump(itemid2item, f)
# del model
# del user_item
# del itemid2item

# user_item = sps.load_npz(os.path.join(model_dir, "user_item.npz"))
# with open(model_file, 'rb') as f:
#     model = pickle.load(f)
# with open(os.path.join(model_dir, 'itemid2item.json'), 'rb') as f:
#     itemid2item = pickle.load(f)

if not debug:
    sps.save_npz(os.path.join(model_dir, "user_item.npz"), user_item)
    with open(model_file, 'wb') as f:
        pickle.dump(model, f)
    with open(os.path.join(model_dir, 'itemid2item.json'), 'wb') as f:
        pickle.dump(itemid2item, f)
    del model
    del user_item
    del itemid2item
    
    user_item = sps.load_npz(os.path.join(model_dir, "user_item.npz"))
    with open(model_file, 'rb') as f:
        model = pickle.load(f)
    with open(os.path.join(model_dir, 'itemid2item.json'), 'rb') as f:
        itemid2item = pickle.load(f)

model_file

model

## Model eval 

# %%time
# train_df = (
#     pl.from_pandas(train_sessions.sample(300))
#         .with_columns(pl.col('user_id').apply(lambda row: get_rec(row, model=model, user_item=user_item
#                                                                  , itemid2item=itemid2item)).alias('next_item_prediction'))
# ).to_pandas()#[['locale', 'next_item_prediction']]


# eval_cols = ['len', 'recall@20', 'recall@100']
# train_df[eval_cols] = train_df.apply(pd_get_recall_at_k, axis=1, result_type='expand')
# print(train_df[eval_cols].mean())

# Model Eval

# train_sessions.head()

# test_sessions.head()

# train_sessions.head()

eval_sessions.head()

# test_sessions.head()

%%time
eval_res = (
    pl.from_pandas(eval_sessions)
        .with_columns(pl.col('user_id').apply(lambda row: get_rec(row, model=model, user_item=user_item
                                                                 , itemid2item=itemid2item)).alias('next_item_prediction'))
).to_pandas()

# eval_res.head()

model_eval(pl.from_pandas(eval_res))

eval_res.to_parquet('../data/als_v1_eval.parquet', engine='pyarrow')

! ls ../data/

%%time
predictions = (
    pl.from_pandas(test_sessions)
        .with_columns(pl.col('user_id').apply(lambda row: get_rec(row, model=model, user_item=user_item
                                                                 , itemid2item=itemid2item)).alias('next_item_prediction'))
).to_pandas()[['locale', 'next_item_prediction']]

# Rec & Save Result 

submit_file

check_predictions(predictions, test_sessions=test_sessions, 
                  # check_products=True, product_df=products
                 )
# Its important that the parquet file you submit is saved with pyarrow backend
predictions.to_parquet(submit_file, engine='pyarrow')

# Submission 

# You can submit with aicrowd-cli, or upload manually on the challenge page.
!aicrowd submission create -c task-1-next-product-recommendation -f {submit_file}

# Example 

# # -*- coding: utf-8 -*-
# """
# Created on Sun Jun 23 22:20:58 2019

# @author: himansh
# """
# #import libraries
# import sys
# import pandas as pd
# import numpy as np
# import scipy.sparse as sparse
# from scipy.sparse.linalg import spsolve
# import random



# from sklearn.preprocessing import MinMaxScaler

# import implicit 
# from datetime import datetime, timedelta


# #Data Preprocessing
# def create_data(datapath,start_date,end_date):
#     df=pd.read_csv(datapath)
#     df=df.assign(date=pd.Series(datetime.fromtimestamp(a/1000).date() for a in df.timestamp))
#     df=df.sort_values(by='date').reset_index(drop=True) # for some reasons RetailRocket did NOT sort data by date
#     df=df[(df.date>=datetime.strptime(start_date,'%Y-%m-%d').date())&(df.date<=datetime.strptime(end_date,'%Y-%m-%d').date())]
#     df=df[['visitorid','itemid','event']]
#     return df


# #Download the kaggle RetailRocket data and give the events.csv file path
# datapath= 'events.csv'
# data=create_data(datapath,'2015-5-3','2015-5-3')
# data['visitorid'] = data['visitorid'].astype("category")
# data['itemid'] = data['itemid'].astype("category")
# data['visitor_id'] = data['visitorid'].cat.codes
# data['item_id'] = data['itemid'].cat.codes

# data['event']=data['event'].astype('category')
# data['event']=data['event'].cat.codes


# sparse_item_user = sparse.csr_matrix((data['event'].astype(float), (data['item_id'], data['visitor_id'])))
# sparse_user_item = sparse.csr_matrix((data['event'].astype(float), (data['visitor_id'], data['item_id'])))





# #Building the model
# model = implicit.als.AlternatingLeastSquares(factors=20, regularization=0.1, iterations=20)


# # ???
# # alpha_val = 40
# # data_conf = (sparse_item_user * alpha_val).astype('double')


# ###USING THE MODEL

# #Get Recommendations
# user_id =   14
# recommended = model.recommend(user_id, 
#                               sparse_user_item[user_id],
#                               filter_already_liked_items=True
#                              )[0].tolist()
# print(recommended)




# # #Get similar items
# # item_id = 7
# # n_similar = 3
# # similar = model.similar_items(item_id, n_similar)
# # print(similar)

# user_item = sps.coo_matrix(
#       (
#           np.ones(data.shape[0]), # We're using a matrix of ones, but using type weights or repurchase weights could help!
#           (data['visitor_id'],
#           data['item_id'])
#       ),
#       dtype='int8'
#     ).tocsr()
# user_item.todense().shape
# new_model = implicit.als.AlternatingLeastSquares(factors=100, regularization=0.1, iterations=20)

# new_model.fit(user_items=user_item, show_progress=True)


# user_id =   14
# recommended = model.recommend(user_id, 
#                               user_item[user_id, :],
#                               filter_already_liked_items=True
#                              )[0].tolist()
# print(recommended)
