==> notebooks/1_task1_word2vector_for_eval.ipynb <==
# Packages

%load_ext autoreload
%autoreload 2
import sys
import logging
base_dir = '../'
sys.path.append(base_dir)
import os
from gensim.similarities.annoy import AnnoyIndexer

from utils import *

import os
import pickle
import numpy as np
import pandas as pd
from gensim.models import Word2Vec
from tqdm import tqdm
import polars as pl
from annoy import AnnoyIndex
import polars as pl
import polars as pl
from utils import *
from src.eval import model_eval
from src.config import (
    raw_data_session_id_dir, candidate_dir, model_for_eval, candidate_file_name, submit_file_name
    # , w2v_model_file_name, w2v_index_file_name
)


# Config

debug = False

topn = 200
model_version = 'w2v_v3'
if debug:
    n_rows = 100
else:
    n_rows = None

    
task = 'task1'
# Common setting
model_dir = f'../model_training/{model_version}/'
# target locales: locales needed for task1
target_locals = ["DE", 'JP', 'UK']
submit_file = os.path.join('../data/sub_files/', 
                           submit_file_name.format(
                                task=task
                               , model_version=model_version
                               , model_for_eval=model_for_eval
                               , topn=topn
                           )
                          )

train_cg_file = os.path.join(base_dir,
                             candidate_dir, 
            candidate_file_name.format(
                task=task
                , data_type='train'
                , model_version=model_version
                , model_for_eval=model_for_eval
                , topn=topn
            )
                            )
eval_cg_file = os.path.join(base_dir,
                            candidate_dir, 
            candidate_file_name.format(
                task=task
                , data_type='eval'
                , model_version=model_version
                , model_for_eval=model_for_eval
                , topn=topn
            )
                            )
test_cg_file = os.path.join(base_dir,
                            candidate_dir, 
            candidate_file_name.format(
                task=task
                , data_type='test'
                , model_version=model_version
                , model_for_eval=model_for_eval
                , topn=topn
            )
                            )

# specific setting
num_tree = 100
# if model_for_eval:
w2v_model_file = os.path.join(model_dir, f"{model_for_eval}.model")
annoy_index_file = os.path.join(model_dir, f"{str(num_tree)}_{model_for_eval}.index")

# submit_file_name

! mkdir {model_dir}

print(f"{w2v_model_file}; \n {annoy_index_file}; \n {submit_file}  \n {train_cg_file} \n {eval_cg_file} \n {test_cg_file}")

# Function 

def w2v_rec(target_df, w2vec, annoy_index, topn=100):
    final_cols = ['session_id', 'next_item_prediction']
    def get_next_items(x):
        prev_items = x['prev_items']
        local_rec = x['next_item_prediction']
        final = [ele for ele in local_rec if ele not in prev_items]
        return final
    target_df = (
        target_df.with_columns(
            pl.col('prev_items').apply(lambda x: 
                                       # x.to_list()
                                       list(map(list, zip(*w2vec.wv.most_similar(positive=x.to_list(),
                                                                     topn=topn+10,
                                                                     indexer=annoy_index))))[0]
                                      ).alias('next_item_prediction'))
        .with_columns(
                    pl.struct(["prev_items", "next_item_prediction"]).apply(
                        lambda x: get_next_items(x)).arr.head(topn).alias('next_item_prediction'))
        .select(
          final_cols
        )
        .with_columns(
            pl.col('next_item_prediction').arr.lengths().alias('rec_num')
        )
    )
    return target_df

# Word2vec

## Train data process

train_pl = pl.scan_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_train.parquet'), n_rows=n_rows).filter(pl.col('locale').is_in(target_locals)).with_columns(pl.col('prev_items').apply(str2list))

eval_pl = pl.scan_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_eval.parquet'), n_rows=n_rows).filter(pl.col('locale').is_in(target_locals)).with_columns(pl.col('prev_items').apply(str2list))

# df_sess.head(3).collect()
test_pl = pl.scan_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_test_task1.parquet'), n_rows=n_rows).with_columns(pl.col('prev_items').apply(str2list))


## Train model & annnoy index 

train_data = train_pl.with_columns(
    pl.col('prev_items').arr.concat(pl.col('next_item')) 
)
if not model_for_eval:
    eval_data = eval_pl.with_columns(
    pl.col('prev_items').arr.concat(pl.col('next_item')) 
)
else:
    eval_data = eval_pl
test_data = test_pl

cols_to_keep = ['prev_items']
sentences = (
    pl.concat([train_data.select(cols_to_keep), eval_data.select(cols_to_keep), test_data.select(cols_to_keep)], how='vertical')
).collect().to_dict()['prev_items'].to_list()

# sentences = map(sentences, list)


len(sentences)

# sentences

# sentences

vector_size = 32
epochs = 10
sg = 1 # 1 for skip-gram
pop_thresh = 0.82415
window = 4

# 

# len(sentences)sentences

from gensim.models.callbacks import CallbackAny2Vec

class callback(CallbackAny2Vec):
    '''Callback to print loss after each epoch.'''

    def __init__(self):
        self.epoch = 0
        self.loss_to_be_subed = 0

    def on_epoch_end(self, model):
        loss = model.get_latest_training_loss()
        loss_now = loss - self.loss_to_be_subed
        self.loss_to_be_subed = loss
        print('Loss after epoch {}: {}'.format(self.epoch, loss_now))
        self.epoch += 1

w2vec = Word2Vec(sentences=sentences, vector_size=vector_size, epochs = epochs, sg=sg,
                 min_count=1, workers=14,
                 window=window,
                  compute_loss=True
              , callbacks=[callback()])

w2v_model_file


# 100 trees are being used in this example
annoy_index = AnnoyIndexer(w2vec, num_tree)

## Save Model 

w2vec.save(w2v_model_file)


annoy_index.save(annoy_index_file)

## Load model 

w2v_model_file

annoy_index_file

w2vec = Word2Vec.load(w2v_model_file)
annoy_index = AnnoyIndexer()
annoy_index.load(annoy_index_file)

## Model Eval

topn

eval_candidate_pl = w2v_rec(target_df=eval_pl,
                   w2vec=w2vec, annoy_index=annoy_index, topn=topn)

eval_candidate_pl.schema

# eval_candidate_pl.head(1).collect()

%%time
eval_candidate_pl = eval_candidate_pl.collect()

eval_candidate_pl.select('rec_num').describe()

# model_eval(target_df=eval_pl.join(eval_candidate_pl.lazy(), how='left', on='session_id'))

%%time
# model_eval(target_df=target_df)
eval_final = (
        eval_pl.join(eval_candidate_pl.lazy(), how='left', on='session_id')
        .lazy()
        .with_columns(
            pl.col('next_item_prediction').cast(pl.List(pl.Utf8))
        )
        .with_columns(
            pl.concat_list([pl.col('next_item'), pl.col('next_item_prediction')]).alias('mrr')
        )
        .with_columns(
            pl.col('mrr').arr.eval(
                pl.arg_where(pl.element()==pl.element().first())
            )
        ).with_columns(
            pl.col('mrr').arr.eval(
                pl.when(pl.element()==0).then(0).otherwise(1/pl.element())
            )
        ).with_columns(
            pl.col('mrr').arr.sum()
            , pl.col('next_item_prediction').arr.head(20).arr.contains(pl.col('next_item')).mean().alias('recall@20')
            , pl.col('next_item_prediction').arr.head(100).arr.contains(pl.col('next_item')).mean().alias('recall@100')
            , pl.col('next_item_prediction').arr.head(topn).arr.contains(pl.col('next_item')).mean().alias('recall@all')


        )
)
final_res = eval_final.select(
        pl.count().alias('total_sessions')
        , pl.col('mrr').mean()
        , pl.col('recall@20').mean()
        , pl.col('recall@100').mean()
        , pl.col('recall@all').mean()

    ).collect()
final_res

## Candidate Saving 

test_candidate_pl  = w2v_rec(target_df=test_pl,
                   w2vec=w2vec, annoy_index=annoy_index, topn=100)
train_candidate_pl  = w2v_rec(target_df=train_pl,
                   w2vec=w2vec, annoy_index=annoy_index, topn=100)

%%time
train_candidate_pl = train_candidate_pl.collect()

%%time
test_candidate_pl = test_candidate_pl.collect()

# train_candidate_pl.head()

### Saving 

eval_candidate_pl.write_parquet(eval_cg_file)

train_candidate_pl.write_parquet(train_cg_file)
test_candidate_pl.write_parquet(test_cg_file)

# reprocess_data = True

# def i2i(df):
#     pl_df = pl.from_dataframe(df)
#     pl_df = (
#         pl_df
#             .with_columns(pl.col('prev_items').apply(lambda row: get_rec(row, annoy_index=annoy_index, topn=100)).alias('next_item_prediction'))
#     )
#     return pl_df

# # %%time
# def pd_get_rec(df, w2vec, annoy_index, topn):
#     next_item_prediction_lst = []
#     for a in tqdm(df.iterrows(), total=len(df)):
#         prev_items = a[1]['prev_items']
#         res = [ele.replace('[', '').replace(']', '').replace('\n', '').replace("'", '').replace(' ', '') for ele in prev_items.split(' ')]
#         # print(type(res))
#         similarity_dic = w2vec.wv.most_similar(positive=res, topn=topn, indexer=annoy_index)
#         res = [item for item, simi in similarity_dic]
#         next_item_prediction_lst.append(res)
#     df['next_item_prediction'] = next_item_prediction_lst
#     return df

# Validate predictions ✅ 😄

predictions = test_pl.join(test_candidate_pl.lazy(), how='left', on='session_id').collect()[['locale', 'next_item_prediction']].to_pandas()

check_predictions(predictions, test_sessions=test_pl.collect().to_pandas(), 
                  # check_products=True, product_df=products
                 )

# Its important that the parquet file you submit is saved with pyarrow backend
predictions.to_parquet(submit_file, engine='pyarrow')

submit_file

## Submit to AIcrowd 🚀

# # You can submit with aicrowd-cli, or upload manually on the challenge page.
# !aicrowd submission create -c task-1-next-product-recommendation -f {submit_file}

# ! ls {file_path}

# eval_pl.head().collect()
