==> notebooks/2_task1_item_and_session_features.ipynb <==
# Package 

%load_ext autoreload
%autoreload 2
import sys
import logging
base_dir = '../'
sys.path.append(base_dir)
import os
import warnings
warnings.simplefilter('ignore')

import pickle
import gc
import re
import polars as pl
from collections import defaultdict, Counter

import numpy as np
import pandas as pd
pd.set_option('display.max_columns', None)
from tqdm.auto import tqdm
import polars as pl
from utils import *
from src.eval import model_eval
from src.config import raw_data_session_id_dir, candidate_dir, model_for_eval

# df_prod = pd.read_csv('data/products_train.csv')
# df_prod

# Config 

# ! ls ../model_training/

debug = False


# model_version = 'popular_v1'

topn = 100
if debug:
    n_rows = 1000
else:
    n_rows = None
# debug_session_num = 100
train_data_dir = '.'
test_data_dir = '.'
task = 'task1'

# model_dir = f'../model_training/{model_version}/'

# target locales: locales needed for task1
target_locals = ["DE", 'JP', 'UK']

# if model_for_eval:
# model_file = os.path.join(model_dir, f'nic_{model_for_eval}_for_eval.parquet')
# submit_file = os.path.join('../data/sub_files/', f'submission_{task}_nic_{model_for_eval}_for_eval.parquet')

# candidate_file.format(
#     task='task1'
#     , data_type='train'
#     , model_version=model_version
#     , model_for_eval=model_for_eval
#     , topn=topn
# )

# Data 

train_pl = pl.scan_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_train.parquet'), n_rows=n_rows).filter(pl.col('locale').is_in(target_locals)).with_columns(pl.col('prev_items').apply(str2list))

eval_pl = pl.scan_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_eval.parquet'), n_rows=n_rows).filter(pl.col('locale').is_in(target_locals)).with_columns(pl.col('prev_items').apply(str2list))

# df_sess.head(3).collect()
test_pl = pl.scan_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_test_task1.parquet'), n_rows=n_rows).with_columns(pl.col('prev_items').apply(str2list))
product_pl = pl.scan_parquet(os.path.join(base_dir, raw_data_session_id_dir, 
                                          'products_train.parquet'))


product_pl.head().collect()

# Session features 

cols_to_keep = ['session_id', 'prev_items', 'locale']
session_pl = pl.concat([
    train_pl.select(cols_to_keep),
    eval_pl.select(cols_to_keep), 
    test_pl.select(cols_to_keep)], how='vertical')
        

session_len = session_pl.select('session_id').collect().shape[0]
session_len

product_cols = ['id', 'locale', 'price', 'brand', 'color', 'size']
session_pl = (
    session_pl.explode('prev_items')
        .join(product_pl.select(product_cols), 
              left_on=['prev_items', 'locale'], 
             right_on=['id', 'locale']
             )
        .groupby(['session_id', 'locale'])
        .agg(
            pl.all()
        )
        .with_columns(
            pl.col('price').arr.mean().alias('mean_price')
            , pl.col('price').arr.get(-1).alias('last_price')
            , pl.col('price').arr.min().alias('min_price')
            , pl.col('price').arr.max().alias('max_price')
        )
)

session_pl = session_pl.collect()

assert session_pl.shape[0] == session_len

session_pl.write_parquet('../data/session_item_features/session_features.parquet')

# Top200 for fallback logics

train_data = train_pl.with_columns(
    pl.col('prev_items').arr.concat(pl.col('next_item')) 
)
if not model_for_eval:
    eval_data = eval_pl.with_columns(
    pl.col('prev_items').arr.concat(pl.col('next_item')) 
)
else:
    eval_data = eval_pl
test_data = test_pl
cols_to_keep = ['prev_items', 'locale']
popular_pl = pl.concat([train_data.select(cols_to_keep), eval_data.select(cols_to_keep), test_data.select(cols_to_keep)], how='vertical')
        

# next_item_df[cols].info()

popular_pl.head().collect()

topn = 200
locale_popular_pl = (
    popular_pl
        .explode('prev_items')
        .groupby(['locale', 'prev_items'])
        .agg(
            pl.count()
        )
        .with_columns(
            pl.col('count').rank(method='ordinal', descending=True).over('locale').alias('rank')
        )
        .filter(pl.col('rank')<=topn)
        .with_columns(
            pl.col('count').max().over('locale').alias('max_count')
            , pl.col('count').min().over('locale').alias('min_count')
        )
        .with_columns(
            ((pl.col('count')-pl.col('min_count'))/(pl.col('max_count')-pl.col('min_count'))).alias('weight')
        )
        .sort('locale', 'rank')
        .select(
            'locale'
            , 'prev_items'
            , 'weight'
        )
        .groupby('locale')
        .agg(
            pl.col('weight').alias('locale_popular_weight')
            , pl.col('prev_items').alias('locale_popular_rec')
        )
        # .count()#.head(3).collect())
        # .collect()
)

locale_popular_pl.collect()

# Model Eval 

def get_next_items(x):
    prev_items = x['prev_items']
    local_rec = x['locale_popular_rec']
    final = [ele for ele in local_rec if ele not in prev_items]
    return final


eval_res = (
    eval_pl
        .join(locale_popular_pl, how='left', on='locale')
        .with_columns(
            pl.col('locale_popular_rec')#.alias('next_item_prediction')
            , pl.struct(["prev_items", "locale_popular_rec"]).apply(
                        lambda x: get_next_items(x)).alias('next_item_prediction')
        )
)

# eval_res.head().collect()

%%time
model_eval(eval_res)

# eval_res.head().collect()

# model_eval(eval_res)

# Save Model

model_dir

popular_pl.collect().write_parquet(os.path.join(model_dir, 'popular_rec.parquet'))
