==> notebooks/3_task1_ranker.ipynb <==
# Packages 

%load_ext autoreload
%autoreload 2
import sys
import logging
base_dir = '../'
sys.path.append(base_dir)
import os
from utils import *

import joblib
import os
import numpy as np
import pandas as pd
from gensim.models import Word2Vec
from gensim.similarities.annoy import AnnoyIndexer


from annoy import AnnoyIndex
import polars as pl
import implicit
from src.eval import model_eval

import scipy.sparse as sps
from utils import str2list
from src.config import raw_data_session_id_dir, candidate_file_name
from lightgbm import LGBMRanker
from lightgbm import early_stopping
from utils import *
# try github 

# Config 

ranker_train_data_dir = '../data/rank_train_data_v2/'
rank_model_version = 'rank_lgbm_v4'

debug = False

if debug:
    SAMPLE_NUM = 1000
else:
    SAMPLE_NUM = None

candidate_path = '../data/candidates/'
model_dir = '../model_training'
# train_data_dir = '.'
# test_data_dir = '.'
task = 'task1'
w2v_model_version = 'w2v_v3'
nic_model_version = 'nic'

rank_model_dir = os.path.join(model_dir, rank_model_version)
model_for_eval = True
w2v_topn=100
nic_topn=100
# PREDS_PER_SESSION = 100

# num_tree = 100



# # target locales: locales needed for task1
target_locals = ["DE", 'JP', 'UK']

# submit_file = f'submission_{task}_ALS.parquet'
num_tree = 100
w2v_model_dir = os.path.join(model_dir, w2v_model_version)
w2v_model_file = os.path.join(w2v_model_dir, f"{model_for_eval}.model")
annoy_index_file = os.path.join(w2v_model_dir, f"{str(num_tree)}_{model_for_eval}.index")


sub_file = f'../data/sub_files/{rank_model_version}.parque'
eval_sub_file = f'../data/sub_files/eval_{rank_model_version}.parque'

rank_model_dir

! mkdir {rank_model_dir}

# Read data


eval_pl = pl.scan_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_eval.parquet'), n_rows=SAMPLE_NUM).filter(pl.col('locale').is_in(target_locals)).with_columns(pl.col('prev_items').apply(str2list))

# df_sess.head(3).collect()
test_pl = pl.scan_parquet(os.path.join(base_dir, raw_data_session_id_dir, f'sessions_test_{task}.parquet'), n_rows=SAMPLE_NUM).with_columns(pl.col('prev_items').apply(str2list))
test4task3_pl = pl.scan_parquet(os.path.join(base_dir, raw_data_session_id_dir, 'sessions_test_task3.parquet'), n_rows=SAMPLE_NUM).filter(pl.col('locale').is_in(target_locals)).with_columns(pl.col('prev_items').apply(str2list))



! ls {ranker_train_data_dir}

train_candidates = pl.scan_parquet(os.path.join(ranker_train_data_dir, 'train/' '*.parquet'), n_rows=SAMPLE_NUM)
# train_candidates = pl.scan_parquet(os.path.join(ranker_train_data_dir, 'train/' 'part_5.parquet'), n_rows=SAMPLE_NUM)

eval_candidates = pl.scan_parquet(os.path.join(ranker_train_data_dir, 'eval.parquet'), n_rows=SAMPLE_NUM)
test_candidates = pl.scan_parquet(os.path.join(ranker_train_data_dir, 'test.parquet'), n_rows=SAMPLE_NUM)


test4task3_candidates = pl.scan_parquet(os.path.join(ranker_train_data_dir, 'test4task3.parquet'), n_rows=SAMPLE_NUM)

# train_candidates.collect().shape

! ls {ranker_train_data_dir}/train

# Ranker 

## Config 

estimator = 1000
early_stop_rounds = 10
eval_frac = 0.1
if debug:
    early_stop_rounds = 5
    
target = 'target'
feature_cols = [
    'next_item_weight'
    , 'next_item_weight_nfi'
    , 'next_item_weight_co_visit'
    , 'locale'
    , 'w2v_weight'
    , 'last_item_similarity'
    # , 'prev_item_similarity'	
    , 'prev_length'
    , 'price_mean_norm'	
    , 'price_last_norm'
    , 'price_min_norm'
    , 'price_max_norm'
]

categorical_feature=['locale']

## Train model

ranker = LGBMRanker(
    objective="lambdarank",
    metric="ndcg",
    boosting_type="gbdt",
    n_estimators=estimator, 
    importance_type='gain',
    eval_at=[3]
)

### eval data for training 

# eval_candidates.collect().shape

# train_candidates.collect().shape

# eval_candidates.head(3).collect()

# train_candidates.head(3).collect()

train_data_source = 'train' # train or 'eval'

if train_data_source == 'eval':
    target_df = eval_candidates
    negative_sampling = 0.1

    cols = ['session_id', 'next_item', 'next_item_prediction', 'target']
    sample_pl = target_df.select(cols)
    pos = sample_pl.filter(pl.col('target')==1)
    neg = (sample_pl
               .filter(pl.col('target')==0)
          )
    neg_original_len = neg.select('session_id').collect().shape[0]
    print(f'Original negative num: {neg_original_len}')
    neg = neg.groupby('session_id').agg(
        pl.all().sample(frac=negative_sampling)
        )
    neg = neg.explode(neg.columns[1:])
    # # print(neg.shape)
    # # print(neg.head().collect())
    sample_pl = pl.concat([pos, neg], how='vertical')
    target_df = sample_pl.join(target_df, how='left', on=cols).collect().sort('session_id').lazy()
    del sample_pl
else:
    target_df = train_candidates
    del train_candidates

### Train 

# target_df.schema

# target_df.select('session_id').head().collect()#.unique().collect()

target_train_candidates = target_df
del target_df

#### More featurs 

session_features = pl.scan_parquet('../data/session_item_features/session_features.parquet')
product_pl = pl.scan_parquet(os.path.join(base_dir, raw_data_session_id_dir, 
                                          'products_train.parquet'))

# session_features.schema

# target_train_candidates

# target_train_candidates.head(3).collect()

def get_additional_feature(data_pl, session_features=session_features, product_pl=product_pl):
    session_join_keys = ['session_id']
    session_fea_cols = ['mean_price', 'last_price', 'min_price', 'max_price']
    data_pl = (
        data_pl.join(session_features.select(session_join_keys+session_fea_cols+['locale']),
                                    how='left', on=session_join_keys, suffix='_str')
            .join(product_pl.select(['id', 'locale', 'price']), how='left', right_on=['id', 'locale']
                 , left_on=['next_item_prediction', 'locale_str'])
            .with_columns(
              pl.col('price').truediv(pl.col('mean_price')).alias('price_mean_norm')
              , pl.col('price').truediv(pl.col('last_price')).alias('price_last_norm')
              , pl.col('price').truediv(pl.col('min_price')).alias('price_min_norm')
              , pl.col('price').truediv(pl.col('max_price')).alias('price_max_norm')
              # , pl.when(pl.col('min_price')==pl.col('max_price')).then(1).otherwise((pl.col('price')-pl.col('min_price'))/(pl.col('min_price')-pl.col('max_price'))).alias('price_norm')
            )
            .select(
                pl.all().exclude(session_fea_cols+['locale_str', 'price'])
            )

    )
    return data_pl

target_train_candidates = get_additional_feature(data_pl=target_train_candidates)

train_session_id = target_train_candidates.select(pl.col('session_id')).unique().collect()

session_num = train_session_id.shape[0]

random_pl = train_session_id.with_columns(
        pl.Series(name='random', values=np.random.uniform(size=session_num))
)

train_candidates = target_train_candidates.join(random_pl.lazy(), how='left', on='session_id')

sampled_train_candidates = train_candidates.filter(pl.col('random')>eval_frac).sort('session_id', descending=False).collect()
sampled_eval_candidates = train_candidates.filter(pl.col('random')<=eval_frac).sort('session_id', descending=False).collect()

print('Train')
print(sampled_train_candidates.shape)
print(sampled_train_candidates.select('target').mean())
print()
print()
print('Eval')
print(sampled_eval_candidates.shape)
print(sampled_eval_candidates.select('target').mean())

# sampled_train_candidates.head()

<!-- no positive items ->  -->

# sampled_train_candidates.filter(pl.col('session_id')==0).sort('target', descending=True)

# sampled_train_candidates.filter(pl.col('session_id')==5).sort('target', descending=True)

# sampled_train_candidates.head()

# sampled_train_candidates.groupby('session_id').agg(pl.count())

session_lengths_train = sampled_train_candidates.groupby('session_id').agg(pl.count()).select('count').to_pandas()['count'].values.tolist()
session_lengths_eval = sampled_eval_candidates.groupby('session_id').agg(pl.count()).select('count').to_pandas()['count'].values.tolist()

pd.DataFrame(session_lengths_train).describe([0.05*i for i in range(20)])

pd.DataFrame(session_lengths_eval).describe([0.05*i for i in range(20)])

ranker = ranker.fit(
    X=sampled_train_candidates[feature_cols].to_pandas(),
    y=sampled_train_candidates[target].to_pandas(),
    group=session_lengths_train,
    eval_set=[(sampled_train_candidates[feature_cols].to_pandas(), sampled_train_candidates[target].to_pandas()),
           (sampled_eval_candidates[feature_cols].to_pandas(), sampled_eval_candidates[target].to_pandas())
             ],
    eval_group=[session_lengths_train,
                session_lengths_eval
               ]
    , categorical_feature=categorical_feature
    , early_stopping_rounds=early_stop_rounds
    # , callbacks=[early_stopping(early_stop_rounds)]
)

print(sampled_train_candidates.estimated_size(unit='gb'))
del sampled_train_candidates

print(sampled_eval_candidates.estimated_size(unit='gb'))
del sampled_eval_candidates

## Save model 

debug

if not debug:
    joblib.dump(
            value=ranker,
            filename=os.path.join(rank_model_dir, 'model.pkl')
    )

if not debug:
    del ranker

## Load Model 

debug

if not debug:
    ranker = joblib.load(os.path.join(rank_model_dir, 'model.pkl'))
ranker

### Importance 

impotant_df = pd.DataFrame(
    {
        'features': ranker.feature_name_,
        'importance': ranker.feature_importances_
    }
).sort_values('importance', ascending=False)
impotant_df

# Eval ranker 

def rank_rec(candidate_pl, feature_cols, ranker, topn=100):

    candidate_pl = candidate_pl.collect()
    print(f"candidate pl shape: {candidate_pl.shape}")
    inference = ranker.predict(candidate_pl[feature_cols].to_pandas())
    test_result = (candidate_pl
         .lazy()
         .with_columns(
             pl.Series(name='predict', values=inference)
         )
         .with_columns(
            pl.col('predict').rank(method='ordinal', descending=True).over('session_id').alias('rank')
         )
         .sort(['session_id', 'rank'])
         .filter(pl.col('rank')<=topn)
         .groupby(['session_id'])
         .agg(
             pl.col('next_item_prediction')
             , pl.col('next_item').unique()#.arr.get(0).cast(pl.Utf8)
         ).with_columns(
             pl.col('next_item').arr.get(0).alias('next_item')
         )
    )
    return test_result


# target_train_candidates.head().collect()

eval_candidates = get_additional_feature(data_pl=eval_candidates)

ranker_eval_pl = rank_rec(candidate_pl=eval_candidates,
                         feature_cols=feature_cols
                         , ranker=ranker
                        # , topn=10
                         )

model_eval(target_df=eval_pl.join(ranker_eval_pl.lazy(), how='left', on='session_id'))

# model_eval(target_df=eval_pl.join(ranker_eval_pl.lazy(), how='left', on='session_id'))

# Test inference 

test_candidates = get_additional_feature(data_pl=test_candidates)

test_candidates = test_candidates.collect()
test_candidates.shape

inference = ranker.predict(test_candidates[feature_cols].to_pandas())

test_result = (test_candidates
     .lazy()
     .with_columns(
         pl.Series(name='predict', values=inference)
     )
     .with_columns(
        pl.col('predict').rank(method='ordinal', descending=True).over('session_id').alias('rank')
     )
     .sort(['session_id', 'rank'])
     .filter(pl.col('rank')<=100)
     .groupby(['session_id'])
     .agg(
         pl.col('next_item_prediction')
     )
)
# test_result.head().collect()

test_pl.schema

predictions = test_pl.join(test_result, how='left', on='session_id').collect()[['locale', 'next_item_prediction']].to_pandas()

predictions.shape

check_predictions(predictions, test_sessions=test_pl.collect().to_pandas(), 
                  # check_products=True, product_df=products
                 )
# Its important that the parquet file you submit is saved with pyarrow backend
if not debug:
    predictions.to_parquet(sub_file, engine='pyarrow')

# test2task3 


topn = 1
test4task3_candidates = get_additional_feature(data_pl=test4task3_candidates)
test4task3_candidates = test4task3_candidates.collect()
inference = ranker.predict(test4task3_candidates[feature_cols].to_pandas())
test4task3_result = (test4task3_candidates
     .lazy()
     .with_columns(
         pl.Series(name='predict', values=inference)
     )
     .with_columns(
        pl.col('predict').rank(method='ordinal', descending=True).over('session_id').alias('rank')
     )
     .sort(['session_id', 'rank'])
     .filter(pl.col('rank')<=topn)
     .groupby(['session_id'])
     .agg(
         pl.col('next_item_prediction')
     )
)
# test4task3_result.head().collect()
predictions = test4task3_pl.join(test4task3_result, how='left', on='session_id').collect()
# [['locale', 'next_item_prediction']].to_pandas()

test4task3_sub_file = f'../data/sub_files/{task}_test4task3_{rank_model_version}.parquet'
print(test4task3_sub_file)
predictions.write_parquet(test4task3_sub_file)
predictions.select('locale').to_series().value_counts()

# Submit result 

# # You can submit with aicrowd-cli, or upload manually on the challenge page.
!aicrowd submission create -c task-1-next-product-recommendation -f {sub_file}
